# Project/IT/backend/emsp/openapi-schema.yml
openapi: 3.0.2
info:
  title: ''
  version: ''
paths:
...
  /api/cs/:
    get:
      operationId: listChargingStations
      description: ''
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: power_source
        required: false
        in: query
        description: power_source
        schema:
          type: string
          enum:
          - DSO
          - Battery
          - Mix
      - name: stations__type
        required: false
        in: query
        description: stations__type
        schema:
          type: string
          enum:
          - Slow
          - Fast
          - Rapid
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 123
                  next:
                    type: string
                    nullable: true
                    format: uri
                    example: http://api.example.org/accounts/?page=4
                  previous:
                    type: string
                    nullable: true
                    format: uri
                    example: http://api.example.org/accounts/?page=2
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChargingStationModel'
          description: ''
      tags:
      - api
    post:
      operationId: createChargingStation
      description: ''
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChargingStationModel'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ChargingStationModel'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ChargingStationModel'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargingStationModel'
          description: ''
      tags:
      - api
...
  /api/cs/{id}/external_status/:
    get:
      operationId: externalStatusChargingStation
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this Charging Station.
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargingStationModel'
          description: ''
      tags:
      - api
...
components:
  schemas:
...
    ChargingStationModel:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        location:
          type: string
          readOnly: true
        power_source:
          enum:
          - DSO
          - Battery
          - Mix
          type: string
        price:
          type: integer
          maximum: 2147483647
          minimum: 0
      required:
      - price
...